feat: Update the jargon lists
I have run the script the script on a large set of Markdown files
I used a lot of different projects, mainly the ones I have worked on
in the past.
I limited myself on projects that were released under MIT license.

7938 Markdown files found
5649027 words found
149650 unique words found
26669 words with at least 2 uppercase letters found

I manually reviewed the list of words and added the ones I found relevant in the
known.txt file, and the ones I found irrelevant in the ignore.txt file.

These lists are not exhaustive, they are the results of days of works.
I still have thousands of words to review.
I will continue to update these lists in the future.

But these lists are already a good starting point to find jargon words.